clean pathable
