+++ app/models/communication/website/agenda/category.rb
@@ -39,7 +39,8 @@ class Communication::Website::Agenda::Category < ApplicationRecord
   include AsDirectObject
   include Contentful
   include Sanitizable
-  include Sluggable
+  include Sluggable # We override slug_unavailable? method
+  include Pathable # Included after Sluggable to make sure slug is correct before anything
   include WithBlobs
   include WithFeaturedImage
   include WithMenuItemTarget
@@ -47,7 +48,7 @@ class Communication::Website::Agenda::Category < ApplicationRecord
   include WithPosition
   include WithTranslations
   include WithUniversity
-  
+
   belongs_to              :parent,
                           class_name: 'Communication::Website::Agenda::Category',
                           optional: true
