+++ app/models/education/program.rb
@@ -60,6 +60,7 @@ class Education::Program < ApplicationRecord
   include Contentful
   include Sanitizable
   include Sluggable
+  include Pathable # Included after Sluggable to make sure slug is correct before anything
   include WebsitesLinkable
   include WithAccessibility
   include WithAlumni
@@ -110,8 +111,6 @@ class Education::Program < ApplicationRecord
 
   validates_presence_of :name
   validates :downloadable_summary, size: { less_than: 50.megabytes }
-
-  after_save :update_children_paths, if: :saved_change_to_path?
 
   scope :published, -> { where(published: true) }
   scope :ordered_by_name, -> { order(:name) }
@@ -166,13 +165,6 @@ class Education::Program < ApplicationRecord
     references = schools + siblings + descendants
     references << parent if parent.present?
     references
-  end
-
-  def update_children_paths
-    children.each do |child|
-      child.update_column :path, child.generated_path
-      child.update_children_paths
-    end
   end
 
   #####################
